push state down
